requirements.txt
pip~=21.1.1
cryptography~=35.0.0
setuptools~=56.0.0
binance-connector
pandas~=1.5
numpy~=1.20.3
requests~=2.26.0
websocket~=0.2.1
python-binance